feat: add create and delete product endpoints

+++ src/domain/use-cases/products/delete-product.use-case.ts
@@ -1,7 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { ProductRepository } from '../../repositories/product.repository';
 
 interface DeleteProductUseCaseRequest {
+  userId: string;
   productId: string;
 }
 
@@ -9,11 +14,18 @@ interface DeleteProductUseCaseRequest {
 export class DeleteProductUseCase {
   constructor(private productRepository: ProductRepository) {}
 
-  async execute({ productId }: DeleteProductUseCaseRequest): Promise<void> {
+  async execute({
+    productId,
+    userId,
+  }: DeleteProductUseCaseRequest): Promise<void> {
     const productExists = await this.productRepository.findById(productId);
 
     if (!productExists) {
       throw new NotFoundException('Resource not found');
+    }
+
+    if (productExists.userId !== userId) {
+      throw new UnauthorizedException('Unauthorized access');
     }
 
     await this.productRepository.delete(productId);
